Feature accomodation fix and auth button comes in the backend now

+++ backend/src/main/java/com/travelease/backend/accommodation/dto/HotelReviewResponse.java
@@ -1,6 +1,5 @@
 package com.travelease.backend.accommodation.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record HotelReviewResponse(
@@ -8,7 +7,7 @@ public record HotelReviewResponse(
         UUID hotelId,
         UUID userId,
         String userName,
-        BigDecimal rating,
+        Integer rating,
         String comment
 ) {
 }
